Color legend for DaparToolshed

Shows a legend based on the tags in the package 'DaparToolshed'

custom_metacell_colors()

colorLegend_ui(id)

colorLegend_server(
  id,
  presentTags = reactive({
     NULL
 }),
  hide.white = TRUE
)

colorLegend(obj = SummarizedExperiment::SummarizedExperiment())

Arguments

id

A character(1) which is the id of the shiny module.

presentTags

A vector of character() which correspond to the tags.

hide.white

A boolean() to indicate whether the white cells must be hidden or not.

obj

An instance of the class SummarizedExperiment.

Value

A vector

NA

NA

A shiny app

Examples

if (FALSE) { # \dontrun{
data(vdata)
shiny::runApp(colorLegend(vdata[[1]]))
} # }