Displays a correlation matrix of the quantitative data of a numeric matrix.

xxxx

Usage

omXplore_density_ui(id)

omXplore_density_server(
  id,
  obj = reactive({
     NULL
 }),
  i = reactive({
     NULL
 }),
  pal.name = reactive({
     NULL
 })
)

densityPlot(data, conds = NULL, pal.name = NULL)

omXplore_density(obj, i)

Arguments

id

xxx

obj

An instance of the class SummarizedExperiment

i

xxx

pal.name

A character(1) which is the name of the palette from the package RColorBrewer from which the colors are taken. Default value is 'Set1'.

data

xxx

conds

xxx

Value

NA

NA

A plot

A shiny app

Examples

if (interactive()) {
  data(vdata)
  omXplore_density(vdata, 1)
}

data(vdata)
qdata <- SummarizedExperiment::assay(vdata[[1]])
conds <- get_group(vdata)
densityPlot(qdata, conds)