Bar plot of missing values per lines using highcharter.

This method plots a bar plot which represents the distribution of the number of missing values (NA) per lines (i.e. proteins).

view_dataset_ui(id)

view_dataset_server(
  id,
  obj = reactive({
     NULL
 }),
  addons = list(),
  useModal = TRUE,
  verbose = FALSE
)

view_dataset(obj = NULL, addons = NULL, useModal = TRUE)

Arguments

id

A character(1) for the 'id' of the shiny module. It must be the same as for the '*_ui' function.

obj

An instance of the class MultiAssayExperiment.

addons

A list to configure the other shiny apps to integrate. Each item correspond to one package:

the name of the slot is the name of the package

the content of the slot is a vector composed of the generic name of the shiny app. Each of the apps listed here must be an exported app of the package. For example, given the value addons = list(testPkg = c('foo', 'foo2')). That means that the package called "testPkg" must provide the four functions: foo1_ui(), foo1_server() and foo2_ui(), foo2_server())

useModal

A boolean(1) that indicates whether to open plot modules in a modal window or not. Default is TRUE.

verbose

A boolean for verbose mode. Default is FALSE.

Value

NA

NA

NA

A shiny application which wraps the functions view_dataset_ui() and the view_dataset_server()

Details

distribution of the missing values per line,

a bar plot which represents the distribution of the number of missing values (NA) per lines (i.e. proteins) and per conditions,

Histogram of missing values.

Variance : Builds a densityplot of the CV of entities in numeric matrix. The CV is calculated for each condition present in the dataset (see the slot 'Condition' in the colData() DataFrame)

Heatmap:

The function heatmapD()

The function [] is inspired from the function 'heatmap.2' that displays a numeric matrix. For more information, please refer to the help of the heatmap.2 function.

Missing values

#' - distribution of the missing values per line,

a bar plot which represents the distribution of the number of missing values (NA) per lines (ie proteins) and per conditions,

Histogram of missing values.

Author

Samuel Wieczorek, Enora Fremy

Examples

if (FALSE) { # \dontrun{
  data(vdata)
  addons <- list(omXplore = c("extFoo1", "extFoo2"))
  runApp(view_dataset(vdata, addons))
  
  shiny::runApp(view_dataset(vdata))
} # }

if (interactive()) {
  data(vdata)
  view_dataset(vdata)
}