Displays different intensity plots.

Displays different intensity plots.

omXplore_intensity_ui(id)

omXplore_intensity_server(
  id,
  obj = reactive({
     NULL
 }),
  i = reactive({
     1
 }),
  track.indices = reactive({
     NULL
 }),
  remoteReset = reactive({
     NULL
 }),
  is.enabled = reactive({
     TRUE
 })
)

omXplore_intensity(obj, i = NULL, withTracking = FALSE)

boxPlot(obj, conds, legend = NULL, pal = NULL, subset = NULL)

violinPlot(data, conds, subset = NULL, pal.name = "Set1")

Arguments

id

A character(1) which is the id of the shiny module.

obj

A instance of the class MultiAssayExperiment

i

An integer which is the index of the assay in the param obj

track.indices

A vector of integers which are the indices of lines to track.

remoteReset

An integer to activate the reset functions

is.enabled

A boolean that indicates whether the widgets should be enabled or disabled. Default is TRUE

withTracking

A boolean(1) indicating whether the tracking option is activated or not.

conds

A vector indicating the name of each sample.

legend

A vector of the conditions (one condition per sample).

pal

A basis palette for the boxes which length must be equal to the number of unique conditions in the dataset.

subset

A integer() vector of index indicating the indices of rows in the dataset to highlight

data

A data.frame() of quantitaive data

pal.name

A character(1) which is the name of the palette from the package RColorBrewer from which the colors are taken. Default value is 'Set1'.

Value

NA

NA

A shiny app

A boxplot

A violin plot

Author

Samuel Wieczorek, Anais Courtier, Enora Fremy

Examples

if (FALSE) { # \dontrun{
  data(vdata)
  shiny::runApp(omXplore_intensity(vdata, 1))
  
data(sub_R25)
conds <- legend <- SummarizedExperiment::colData(sub_R25)$group
pal <- ExtendPalette(length(unique(conds)))
boxPlot(sub_R25[[1]], conds, legend, pal, seq_len(10))

shiny::runApp(omXplore_intensity(sub_R25, 1, withTracking = TRUE))
} # }