Shiny modules used by omXplore

These functions are relative to external modules that can be added into omXplore UI:

listShinyApps(): Show the shiny modules recognized by omXplore and ready to bu integrated in the UI of the function view_dataset()

listPlotModules(): Show the shiny modules function names (only prefixes) recognized by omXplore and ready to use in the UI.

addModules(): Add external shiny module(s) to the R global environment in such a way (specific prefix renaming of the functions) that it can be discovered by the function view_dataset() of the package omXplore during its launch.

Usage

addModules(addons = list())

listShinyApps(location = "both")

listPlotModules(location = "both")

Arguments

addons

A list in which each item:

is named by the name of a package containing the modules to add,

contains the name of the shiny modules to integrate (without '_ui' nor '_server' suffixes)

location

A character(0) to indicate which modules to list. Available values are: 'builtin', 'external' and 'both' (default).

Value

NA

A vector

A vector

Examples

listShinyApps()
#> [1] "omXplore_cc()"          "omXplore_corrmatrix()"  "omXplore_density()"    
#> [4] "omXplore_heatmap()"     "omXplore_intensity()"   "omXplore_pca()"        
#> [7] "omXplore_tabExplorer()" "omXplore_variance()"   
listPlotModules()
#> [1] "omXplore_cc"          "omXplore_corrmatrix"  "omXplore_density"    
#> [4] "omXplore_heatmap"     "omXplore_intensity"   "omXplore_pca"        
#> [7] "omXplore_tabExplorer" "omXplore_variance"   

#####################################################
# Integration of a module in the package 'mypackage'
#####################################################
# \donttest{
addons <- list(omXplore = c("extFoo1", "extFoo2"))
addModules(addons)
# }