Displays a correlation matrix of the quantitative data of a numeric matrix.

This function is a wrapper to heatmap.2() that displays assay data in an instance of SummarizedExperiment. For more details, see heatmap.2().

omXplore_heatmap_ui(id)

omXplore_heatmap_server(
  id,
  obj = reactive({
     NULL
 }),
  i = reactive({
     NULL
 })
)

omXplore_heatmap(obj, i)

heatmapD(
  qdata,
  conds,
  distance = "euclidean",
  cluster = "complete",
  dendro = FALSE
)

mv.heatmap(
  x,
  col = grDevices::heat.colors(100),
  srtCol = NULL,
  labCol = NULL,
  labRow = NULL,
  key = TRUE,
  key.title = NULL,
  main = NULL,
  ylab = NULL
)

heatmapForMissingValues(
  x,
  col = NULL,
  srtCol = NULL,
  labCol = NULL,
  labRow = NULL,
  key = TRUE,
  key.title = NULL,
  main = NULL,
  ylab = NULL
)

Arguments

id

A character(1) which is the id of the shiny module.

obj

An instance of a class MultiAssayExperiment.

i

An integer which is the index of the assay in the param obj

qdata

A data.frame() of quantitative data.

conds

A vector indicating the name of each sample.

distance

The distance used by the clustering algorithm to compute the dendrogram.

cluster

the clustering algorithm used to build the dendrogram.

dendro

A boolean to indicate fi the dendrogram has to be displayed

x

A matrix or array containing the quantitative data.

col

Colors used for the image. Defaults to heat colors (heat.colors).

srtCol

Angle of column conds, in degrees from horizontal

labCol

Character vectors with column conds to use.

labRow

Character vectors with row conds to use.

key

Logical indicating whether a color-key should be shown.

key.title

Main title of the color key. If set to NA no title will be plotted.

main

Main title; default to none.

ylab

y-axis title; default to none.

Value

NA

NA

A shiny app

A heatmap

A heatmap

A heatmap

Author

Florence Combes, Samuel Wieczorek, Enora Fremy

Examples

if (FALSE) { # \dontrun{
  data(vdata)
  omXplore_heatmap(vdata, 1)
} # }