Display a CC

xxxx

Usage

buildGraph(cc = NULL, meta = NULL)

display.CC.visNet(
  g = NULL,
  layout = "layout_with_fr",
  obj = NULL,
  prot.tooltip = NULL,
  pept.tooltip = NULL
)

plotCCJitter(df, clickFunction = NULL)

GetCCInfos(cc)

Arguments

cc

xxx

meta

xxx

g

A cc (a list)

layout

xxxxx

obj

xxx

prot.tooltip

xxx

pept.tooltip

xxx

df

xxxx

clickFunction

xxxx

Value

A list

A plot

A plot

A list of three items:

One_One: the number of cc composed of one protein and one peptide

One_Multi: the number of cc composed of one protein and several peptides

Multi_Multi: the number of cc composed of several proteins and several (shared) peptides.

Author

Thomas Burger, Samuel Wieczorek

Examples

data(sub_R25)
se <- sub_R25[[1]]
g <- buildGraph(get_cc(se)[[1]])
display.CC.visNet(g)

data(sub_R25)
GetCCInfos(get_cc(sub_R25[[1]]))
#> $One_One
#> $One_One[[1]]
#> 1 x 1 sparse Matrix of class "dgCMatrix"
#>     1017
#> 161    1
#> 
#> $One_One[[2]]
#> 1 x 1 sparse Matrix of class "dgCMatrix"
#>     1206
#> 157    1
#> 
#> $One_One[[3]]
#> 1 x 1 sparse Matrix of class "dgCMatrix"
#>     1262
#> 159    1
#> 
#> $One_One[[4]]
#> 1 x 1 sparse Matrix of class "dgCMatrix"
#>     1574
#> 169    1
#> 
#> $One_One[[5]]
#> 1 x 1 sparse Matrix of class "dgCMatrix"
#>     1652
#> 154    1
#> 
#> $One_One[[6]]
#> 1 x 1 sparse Matrix of class "dgCMatrix"
#>     1819
#> 151    1
#> 
#> $One_One[[7]]
#> 1 x 1 sparse Matrix of class "dgCMatrix"
#>     1969
#> 158    1
#> 
#> $One_One[[8]]
#> 1 x 1 sparse Matrix of class "dgCMatrix"
#>     238
#> 155   1
#> 
#> $One_One[[9]]
#> 1 x 1 sparse Matrix of class "dgCMatrix"
#>     3
#> 167 1
#> 
#> $One_One[[10]]
#> 1 x 1 sparse Matrix of class "dgCMatrix"
#>     30
#> 156  1
#> 
#> $One_One[[11]]
#> 1 x 1 sparse Matrix of class "dgCMatrix"
#>     379
#> 165   1
#> 
#> $One_One[[12]]
#> 1 x 1 sparse Matrix of class "dgCMatrix"
#>     498
#> 164   1
#> 
#> $One_One[[13]]
#> 1 x 1 sparse Matrix of class "dgCMatrix"
#>     661
#> 160   1
#> 
#> $One_One[[14]]
#> 1 x 1 sparse Matrix of class "dgCMatrix"
#>     723
#> 150   1
#> 
#> $One_One[[15]]
#> 1 x 1 sparse Matrix of class "dgCMatrix"
#>     826
#> 168   1
#> 
#> $One_One[[16]]
#> 1 x 1 sparse Matrix of class "dgCMatrix"
#>     829
#> 149   1
#> 
#> 
#> $One_Multi
#> $One_Multi[[1]]
#> 2 x 1 sparse Matrix of class "dgCMatrix"
#>     120
#> 152   1
#> 153   1
#> 
#> 
#> $Multi_Multi
#> $Multi_Multi[[1]]
#> 2 x 2 sparse Matrix of class "dgCMatrix"
#>     1578 1577
#> 162    1    1
#> 163    1    .
#> 
#> $Multi_Multi[[2]]
#> 1 x 2 sparse Matrix of class "dgCMatrix"
#>     1685 2260
#> 166    1    1
#> 
#>