Quantitative metadata vocabulary for entities

This function gives the vocabulary used for the quantitative metadata of each entity in each condition.

metacell.def(level)

Parent(level, node = NULL)

Children(level, parent = NULL)

GetMetacellTags(metacells = NULL, level = NULL, onlyPresent = FALSE)

Arguments

level

A string corresponding to the type of object

node

The name of the node for which one wants its parent

#' @examples Parent('protein', 'Missing') Parent('protein', 'Missing POV') Parent('protein', c('Missing POV', 'Missing MEC')) Parent('protein', c('Missing', 'Missing POV', 'Missing MEC'))

parent

The name og the parent node

metacells

A data.frame() representing the cell metadata

onlyPresent

A boolean(1)

Value

A data.frame containing the different tags and corresponding colors for the level given in parameter

A list

A vector

A vector

A vector

Glossary

Peptide-level vocabulary

|– 'Any' | | | |– 1.0 'Quantified' | | | | | |– 1.1 "Quant. by direct id" (color 4, white) | | | | | |– 1.2 "Quant. by recovery" (color 3, lightgrey) | | | |– 2.0 "Missing" (no color) | | | | | |– 2.1 "Missing POV" (color 1) | | | | | |– 2.2 'Missing MEC' (color 2) | | | |– 3.0 'Imputed' | | | | | |– 3.1 'Imputed POV' (color 1) | | | | | |– 3.2 'Imputed MEC' (color 2)

Protein-level vocabulary: |– 'Any' | | | |– 1.0 'Quantified' | | | | | |– 1.1 "Quant. by direct id" (color 4, white) | | | | | |– 1.2 "Quant. by recovery" (color 3, lightgrey) | | | |– 2.0 "Missing" | | | | | |– 2.1 "Missing POV" (color 1) | | | | | |– 2.2 'Missing MEC' (color 2) | | | |– 3.0 'Imputed' | | | | | |– 3.1 'Imputed POV' (color 1) | | | | | |– 3.2 'Imputed MEC' (color 2) | | | |– 4.0 'Combined tags' (color 3bis, lightgrey)

Conversion to the glossary

A generic conversion

Conversion for Proline datasets

Conversion from Maxquant datasets

Author

Thomas Burger, Samuel Wieczorek

Samuel Wieczorek

Examples

metacell.def("protein")
#>                   node     parent   color
#> 1                  Any              white
#> 2           Quantified        Any #0A31D0
#> 3  Quant. by direct id Quantified #6178D9
#> 4   Quant. by recovery Quantified #B9C4F2
#> 5              Missing        Any #CF8205
#> 6          Missing POV    Missing #E5A947
#> 7          Missing MEC    Missing #F1CA8A
#> 8              Imputed        Any #A40C0C
#> 9          Imputed POV    Imputed #E34343
#> 10         Imputed MEC    Imputed #F59898
#> 11       Combined tags        Any #1E8E05
metacell.def("peptide")
#>                   node     parent   color
#> 1                  Any              white
#> 2           Quantified        Any #0A31D0
#> 3  Quant. by direct id Quantified #6178D9
#> 4   Quant. by recovery Quantified #B9C4F2
#> 5              Missing        Any #CF8205
#> 6          Missing POV    Missing #E5A947
#> 7          Missing MEC    Missing #F1CA8A
#> 8              Imputed        Any #A40C0C
#> 9          Imputed POV    Imputed #E34343
#> 10         Imputed MEC    Imputed #F59898

#-----------------------------------------------
# A shiny app to view color legends
#-----------------------------------------------
if(interactive()) {
  data(vdata)
  ui <- qMetacellLegend_ui("legend")

  server <- function(input, output, session) {
    qMetacellLegend_server("legend",
      object = reactive({vdata[[1]]})
    )
  }

  shinyApp(ui = ui, server = server)
}

Children("protein", "Missing")
#> [1] "Missing POV" "Missing MEC"
Children("protein", "Missing POV")
#> character(0)
Children("protein", c("Missing POV", "Missing MEC"))
#> character(0)
Children("protein", c("Missing", "Missing POV", "Missing MEC"))
#> [1] "Missing POV" "Missing MEC"
data(vdata)
metacells <- get_metacell(vdata[[1]])
level <- get_type(vdata[[1]])
GetMetacellTags(metacells, level)
#>  [1] "Quantified"          "Quant. by direct id" "Quant. by recovery" 
#>  [4] "Missing"             "Missing POV"         "Missing MEC"        
#>  [7] "Imputed"             "Imputed POV"         "Imputed MEC"        
#> [10] "Combined tags"